5512: Fix cypress tests; Fix live validation for step 1 and 2;
