+++ cypress/local-only/integration/fileAPetitionUpdated/file-a-petition-step-1-other--donor-transferee-deceased-spouse.cy.ts
@@ -85,19 +85,11 @@ const DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID = [
   'phone-error-message',
 ];
 
-describe('File a petition: Step 2 - Petitioner Information', () => {
+describe('File a petition: Step 1 - Petitioner Information', () => {
   beforeEach(() => {
     loginAsPetitioner();
     cy.visit('/file-a-petition/new');
-
-    cy.get('[data-testid="petition-reason--1"]').focus();
-    cy.get('[data-testid="petition-reason--1"]').type('REASON 1');
-
-    cy.get('[data-testid="petition-fact--1"]').focus();
-    cy.get('[data-testid="petition-fact--1"]').type('FACT 1');
-
-    cy.get('[data-testid="step-1-next-button"]').click();
-    cy.get('[data-testid="step-indicator-current-step-2-icon"]');
+    cy.get('[data-testid="step-indicator-current-step-1-icon"]');
   });
 
   describe('Other', () => {
@@ -118,7 +110,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             cy.get(`[data-testid="${selector}"]`).should('not.exist');
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
+          cy.get('[data-testid="step-1-next-button"]').click();
 
           DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID.forEach((selector: string) => {
             cy.get(`[data-testid="${selector}"]`).should('exist');
@@ -135,8 +127,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             }
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
         it('should allow user to go to step 3 if everything is filled out correctly', () => {
           DOMESTIC_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -150,8 +142,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${input}"]`).type(inputValue);
             }
           });
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
       });
       describe('International', () => {
@@ -165,7 +157,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             },
           );
 
-          cy.get('[data-testid="step-2-next-button"]').click();
+          cy.get('[data-testid="step-1-next-button"]').click();
 
           INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID.forEach(
             (selector: string) => {
@@ -184,8 +176,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             }
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
         it('should allow user to go to step 3 if everything is filled out correctly', () => {
           INTERNATIONAL_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -199,8 +191,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${input}"]`).type(inputValue);
             }
           });
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
       });
     });
@@ -218,7 +210,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             cy.get(`[data-testid="${selector}"]`).should('not.exist');
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
+          cy.get('[data-testid="step-1-next-button"]').click();
 
           DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID.forEach((selector: string) => {
             cy.get(`[data-testid="${selector}"]`).should('exist');
@@ -235,8 +227,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             }
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
         it('should allow user to go to step 3 if everything is filled out correctly', () => {
           DOMESTIC_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -250,8 +242,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${input}"]`).type(inputValue);
             }
           });
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
       });
       describe('International', () => {
@@ -265,7 +257,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             },
           );
 
-          cy.get('[data-testid="step-2-next-button"]').click();
+          cy.get('[data-testid="step-1-next-button"]').click();
 
           INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID.forEach(
             (selector: string) => {
@@ -284,8 +276,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             }
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
         it('should allow user to go to step 3 if everything is filled out correctly', () => {
           INTERNATIONAL_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -299,8 +291,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${input}"]`).type(inputValue);
             }
           });
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
       });
     });
@@ -321,7 +313,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             cy.get(`[data-testid="${selector}"]`).should('not.exist');
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
+          cy.get('[data-testid="step-1-next-button"]').click();
 
           [
             ...DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID,
@@ -348,8 +340,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             }
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
         it('should allow user to go to step 3 if everything is filled out correctly', () => {
           [
@@ -370,8 +362,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${input}"]`).type(inputValue);
             }
           });
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
       });
       describe('International', () => {
@@ -386,7 +378,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             cy.get(`[data-testid="${selector}"]`).should('not.exist');
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
+          cy.get('[data-testid="step-1-next-button"]').click();
 
           [
             ...INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID,
@@ -413,8 +405,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
             }
           });
 
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
         it('should allow user to go to step 3 if everything is filled out correctly', () => {
           [
@@ -435,8 +427,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${input}"]`).type(inputValue);
             }
           });
-          cy.get('[data-testid="step-2-next-button"]').click();
-          cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+          cy.get('[data-testid="step-1-next-button"]').click();
+          cy.get('[data-testid="step-indicator-current-step-2-icon"]');
         });
       });
     });
