+++ cypress/local-only/integration/fileAPetitionUpdated/file-a-petition-step-1-other--minor-or-legally-incompetent.cy.ts
@@ -97,19 +97,11 @@ const DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID = [
   'phone-error-message',
 ];
 
-describe('File a petition: Step 2 - Petitioner Information', () => {
+describe('File a petition: Step 1 - Petitioner Information', () => {
   beforeEach(() => {
     loginAsPetitioner();
     cy.visit('/file-a-petition/new');
-
-    cy.get('[data-testid="petition-reason--1"]').focus();
-    cy.get('[data-testid="petition-reason--1"]').type('REASON 1');
-
-    cy.get('[data-testid="petition-fact--1"]').focus();
-    cy.get('[data-testid="petition-fact--1"]').type('FACT 1');
-
-    cy.get('[data-testid="step-1-next-button"]').click();
-    cy.get('[data-testid="step-indicator-current-step-2-icon"]');
+    cy.get('[data-testid="step-indicator-current-step-1-icon"]');
   });
 
   describe('Other', () => {
@@ -146,7 +138,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
           'not.exist',
         );
 
-        cy.get('[data-testid="step-2-next-button"]').click();
+        cy.get('[data-testid="step-1-next-button"]').click();
 
         cy.get('[data-testid="minor-incompetent-type-error-message"]').should(
           'exist',
@@ -168,7 +160,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${selector}"]`).should('not.exist');
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID.forEach((selector: string) => {
               cy.get(`[data-testid="${selector}"]`).should('exist');
@@ -185,8 +177,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             DOMESTIC_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -200,8 +192,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
         describe('International', () => {
@@ -215,7 +207,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               },
             );
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID.forEach(
               (selector: string) => {
@@ -234,8 +226,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             INTERNATIONAL_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -249,8 +241,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
       });
@@ -270,7 +262,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${selector}"]`).should('not.exist');
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID.forEach((selector: string) => {
               cy.get(`[data-testid="${selector}"]`).should('exist');
@@ -287,8 +279,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             DOMESTIC_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -302,8 +294,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
         describe('International', () => {
@@ -318,7 +310,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               },
             );
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID.forEach(
               (selector: string) => {
@@ -337,8 +329,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             INTERNATIONAL_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -352,8 +344,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
       });
@@ -373,7 +365,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${selector}"]`).should('not.exist');
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID.forEach((selector: string) => {
               cy.get(`[data-testid="${selector}"]`).should('exist');
@@ -390,8 +382,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             DOMESTIC_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -405,8 +397,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
         describe('International', () => {
@@ -420,7 +412,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               },
             );
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID.forEach(
               (selector: string) => {
@@ -439,8 +431,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             INTERNATIONAL_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -454,8 +446,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
       });
@@ -475,7 +467,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${selector}"]`).should('not.exist');
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID.forEach((selector: string) => {
               cy.get(`[data-testid="${selector}"]`).should('exist');
@@ -492,8 +484,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             DOMESTIC_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -507,8 +499,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
         describe('International', () => {
@@ -522,7 +514,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               },
             );
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID.forEach(
               (selector: string) => {
@@ -541,8 +533,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             INTERNATIONAL_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -556,8 +548,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
       });
@@ -577,7 +569,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               cy.get(`[data-testid="${selector}"]`).should('not.exist');
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             DOMESTIC_ERROR_MESSAGES_DATA_TEST_ID.forEach((selector: string) => {
               cy.get(`[data-testid="${selector}"]`).should('exist');
@@ -594,8 +586,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             DOMESTIC_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -609,8 +601,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
         describe('International', () => {
@@ -624,7 +616,7 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               },
             );
 
-            cy.get('[data-testid="step-2-next-button"]').click();
+            cy.get('[data-testid="step-1-next-button"]').click();
 
             INTERNATIONAL_ERROR_MESSAGES_DATA_TEST_ID.forEach(
               (selector: string) => {
@@ -643,8 +635,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
               }
             });
 
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
           it('should allow user to go to step 3 if everything is filled out correctly', () => {
             INTERNATIONAL_FORM_INPUT_DATA.forEach(inputInfo => {
@@ -658,8 +650,8 @@ describe('File a petition: Step 2 - Petitioner Information', () => {
                 cy.get(`[data-testid="${input}"]`).type(inputValue);
               }
             });
-            cy.get('[data-testid="step-2-next-button"]').click();
-            cy.get('[data-testid="step-indicator-current-step-3-icon"]');
+            cy.get('[data-testid="step-1-next-button"]').click();
+            cy.get('[data-testid="step-indicator-current-step-2-icon"]');
           });
         });
       });
